+++ blocklets/ai-studio/src/components/multi-tenant-restriction/plan-upgrade.tsx
@@ -3,6 +3,7 @@ import { Icon } from '@iconify-icon/react';
 import ArrowUpIcon from '@iconify-icons/tabler/circle-arrow-up';
 import DiamondIcon from '@iconify-icons/tabler/diamond';
 import InfoCircleIcon from '@iconify-icons/tabler/info-circle';
+import ReceiptIcon from '@iconify-icons/tabler/receipt';
 import { Close } from '@mui/icons-material';
 import {
   Alert,
@@ -19,19 +20,45 @@ import {
   useMediaQuery,
 } from '@mui/material';
 import { useState } from 'react';
+import { joinURL } from 'ufo';
 
+import { PAYMENT_KIT_MOUNT_POINT } from '../../libs/constants';
 import { PricingTable } from './pricing-table';
-import { premiumPlanEnabled, useMultiTenantRestriction, usePlans } from './state';
+import { premiumPlanEnabled, useIsPremiumUser, useMultiTenantRestriction, usePlans } from './state';
+
+const billingLink = joinURL(PAYMENT_KIT_MOUNT_POINT, '/customer');
 
 export function PlanUpgrade() {
   const { hidePlanUpgrade, planUpgradeVisible, type } = useMultiTenantRestriction();
   const { t, locale } = useLocaleContext();
   const downSm = useMediaQuery<Theme>((theme) => theme.breakpoints.down('sm'));
   const [billingCycle, setBillingCycle] = useState<'monthly' | 'yearly'>('monthly');
+  const isPremiumUser = useIsPremiumUser();
   const plans = usePlans();
 
   if (!plans || !premiumPlanEnabled) {
     return null;
+  }
+
+  if (isPremiumUser) {
+    plans[1]!.billingLink = (
+      <Box
+        component={Link}
+        href={billingLink}
+        target="_blank"
+        sx={{
+          alignSelf: 'flex-start',
+          display: 'flex',
+          alignItems: 'center',
+          gap: 0.5,
+          fontSize: 13,
+          color: 'text.secondary',
+          textDecoration: 'underline',
+        }}>
+        <Box component={Icon} icon={ReceiptIcon} sx={{ fontSize: 16 }} />
+        View my billing information
+      </Box>
+    );
   }
 
   return (