+++ packages/ai-runtime/src/front/contexts/Runtime.tsx
@@ -15,11 +15,13 @@ export const RuntimeProvider = ({
   children,
   aid,
   working,
+  debug,
   ApiProps,
 }: {
   children?: ReactNode;
   aid: string;
   working?: boolean;
+  debug?: boolean;
   ApiProps?: Partial<AIGNEApiContextValue>;
 }) => {
   const projectId = useMemo(() => parseIdentity(aid, { rejectWhenError: true }).projectId, [aid]);
@@ -28,6 +30,7 @@ export const RuntimeProvider = ({
     <EntryAgentProvider aid={aid} working={working}>
       <AIGNEApiProvider
         working={({ aid }) => (parseIdentity(aid, { rejectWhenError: true }).projectId === projectId ? working : false)}
+        debug={debug}
         {...ApiProps}>
         <ThemeProvider>
           <RuntimeLocaleProvider>